Various personalisations
I was previously using a fork of sellout/emacs-color-theme-solarized which made
the Solarized colour variables accessible from outside the theme, and provided a
hook so I could set variables and faces to Solarized values at arbitrary points
in my Emacs files.

This was convoluted and ineffective, so I'm updating my fork of this, and
putting all of my Solarized-related customisations in the theme. Changes:

- Update some of the face definitions for my own preferences.

- Add some variable definitions to the theme. This is useful for setting
variables that expect colour values, eg. fci-rule-color.

- Add some utility functions - eg. to reload the Solarized theme in a particular
mode, and to toggle comment visibility.